Separate module for greetservice api
