+++ greetservice-api/src/main/java/com/example/greetservice/GreetService.java
@@ -1,4 +1,4 @@
-package com.example.starter.base;
+package com.example.greetservice;
 
 public interface GreetService {
     String greet(String name);
