+++ greetservice-impl/pom.xml
@@ -14,7 +14,7 @@
     <dependencies>
         <dependency>
             <groupId>com.example</groupId>
-            <artifactId>project-base-karaf</artifactId>
+            <artifactId>greetservice-api</artifactId>
             <version>1.0.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
